Model data classes added
