+++ app/src/main/java/com/example/valorantagents/data/network/model/Ability.kt
@@ -1,0 +1,8 @@
+package com.example.valorantagents.data.network.model
+
+data class Ability (
+    val slot: Slot,
+    val displayName: String,
+    val description: String,
+    val displayIcon: String? = null
+)
