+++ app/src/main/java/com/example/valorantagents/data/network/model/Resource.kt
@@ -14,7 +14,4 @@ sealed class Resource<out T> {
     ) : Resource<Nothing>()
 
     object Loading : Resource<Nothing>()
-
-    object Waiting : Resource<Nothing>()
-
 }
