+++ app/src/main/java/com/example/valorantagents/data/network/model/Role.kt
@@ -1,0 +1,9 @@
+package com.example.valorantagents.data.network.model
+
+data class Role (
+    val uuid: String,
+    val displayName: DisplayName,
+    val description: String,
+    val displayIcon: String,
+    val assetPath: String
+)
