+++ app/src/main/java/com/example/valorantagents/data/network/model/Slot.kt
@@ -1,0 +1,9 @@
+package com.example.valorantagents.data.network.model
+
+enum class Slot {
+    Ability1,
+    Ability2,
+    Grenade,
+    Passive,
+    Ultimate
+}
